refactor(#3037): relocate agent JSON serialization helpers to services layer (backflow 33→31)
Bucket 4 (domain serialization misplaced in server/dto). The agent JSON
builder functions (agent_office_json, agent_skill_json,
dispatched_session_json, timeline_event_json, transcript_json,
dedup_dispatched_sessions, build_channel_deeplinks) and their backing
*View structs lived in crate::server::dto::agents but were consumed only
by the services layer (services::agents::{query,turn}), forcing a
cross-layer backflow.

Move them verbatim (behavior-unchanged) to a new
services::agents::serialization module beside their only callers. The
serde *response container* DTOs (Agent*Response) stay in
server::dto::agents because they are presentation types used solely by
the route handlers; no facade re-export is needed since no server route
consumed the relocated functions.

- backflow 33→31 (services/agents/{query,turn}.rs entries removed)
- gates: cargo fmt --check clean; cargo check --lib clean (0 new warnings)
  AND cargo check --lib --tests clean; serialization + dto tests pass;
  docs freshness passed; ci-script-checks exit 0
- ARCHITECTURE.md regenerated (new module listed); no change-surfaces edit
