+++ src/services/agents/mod.rs
@@ -1,2 +1,3 @@
 pub mod query;
+pub mod serialization;
 pub mod turn;
